docs: 토큰 저장 방법 표기
