--- src/main/java/com/project/scenepickbe/config/SwaggerConfig.java
@@ -1,26 +1,0 @@
-package com.project.scenepickbe.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-
-@Configuration
-public class SwaggerConfig {
-
-	@Bean
-	public OpenAPI openAPI() {
-		return new OpenAPI()
-			.components(new Components())
-			.info(apiInfo());
-	}
-
-	private Info apiInfo() {
-		return new Info()
-			.title("ScenePick API Documentation")
-			.description("ScenePick 프로젝트의 API 명세서입니다.")
-			.version("1.0.0");
-	}
-}
